+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added new option "Visible & Readonly, Disabled always" for field "Visibility and default state" in the GTM template.
 
 ## [1.2.0] - 2025-01-10
 ### Added